+++ data/js/platforms/youtube.js
@@ -258,6 +258,7 @@ let youtube = {
 					}
 				}
 			}
+			return streamData;
 		}/*,
 	"importStreamWebsites":
 		function(id, data, streamListSetting, pageNumber){